Remove example
The existing example contains integration tests for simple v1 payjoins.

V1 payjoin is being removed from the library, and the example is going
to be replaced by what now lives in payjoin-flutter-demo to replace the
integration tests.

BDK and bitcoind integration tests are duplicating the smoke-test happy
path testing while also testing that those libraries contain sufficient
information to satisfy payjoin. Assuming they're typical wallets, this
is already tested by rust-payjoin and the bdk implementations like
mutiny and bitmask-core, so that's redundant and merely toil to maintain.
The final commit in this PR will contain the payjoin-flutter-demo in
order to smoke test the result.