Remove V1 receiver
There is no reason to support the v1 receiver in Dart when v2 is
available. If someone really wants to run a v1 server, they can
use the Rust library. Reducing our API surface makes for less
maintenance and fewer opportunities to introduce bugs.

Header is removed because it is v1 only.

Dart clients using v2 are still backwards compatible with v1.