Support payjoin-ffi-0.21.0
The typestate machines in send and receive receive significant changes
to support the new wire protocol.

The Script type moves to bitcoin_ffi.rs since it requires an impl block
beyond a mere record with fields. The other types can stay in types.rs
for now.

enable-danger-local-https is renamed _danger-local-https in line with
the upstream feature, though we won't make use of this feature until
later integration tests are supported by payjoin-ffi fixtures. For now
we will use production servers to smoke test.