Expose de/serialize for Sender, Receiver
These structs need to be serialized and deserialized in order to be
persisted in a downstream database.

We use json to do this only because that decision was made in
payjoin-cli. A binary encoding could be used as well, but that bikeshed
is to be built at a later date.