Add README notes on Flutter and FRB
The build runner is an important step in flutter work flow that is easily overlooked, good to make note of it in any flutter repo.

Flutter Rust Bridge has some quirks including a sort of circular dependency in the code gen process, make note of it.

Clean up vestigial notes on integration tests which were removed